refactor: Move tests back to their intended location.
