+++ tests/serverpod_test_server/test_integration/auth_module/email_auth_provider/custom_password_hash_test.dart
@@ -1,7 +1,7 @@
 import 'package:serverpod_auth_server/module.dart';
 import 'package:test/test.dart';
 
-import '../../serverpod_test_tools.dart';
+import '../../test_tools/serverpod_test_tools.dart';
 
 void main() async {
   AuthConfig.set(
